criacao e modularizacao de componentes
