+++ src/components/layout.module.css
@@ -3,13 +3,3 @@
   padding: 0 1rem;
   margin: 3rem auto 6rem;
 }
-
-.header {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-.backToHome {
-  margin: 3rem 0 0;
-}
