+++ src/components/profile-header/profile-header.module.css
@@ -1,0 +1,28 @@
+.header {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+}
+
+.borderCircle {
+    border-radius: 9999px;
+    object-fit: fill;
+}
+
+.heading2Xl {
+    font-size: 2.5rem;
+    line-height: 1.2;
+    font-weight: 800;
+    letter-spacing: -0.05rem;
+    margin: 1rem 0;
+}
+
+.headingLg {
+    font-size: 1.5rem;
+    line-height: 1.4;
+    margin: 1rem 0;
+}
+
+.colorInherit {
+    color: inherit;
+}
