+++ src/components/profile-info/profile-info.js
@@ -1,0 +1,15 @@
+import styles from "./profile-info.module.css";
+
+export default function ProfileInfo() {
+  return (
+    <section className={styles.headingMd}>
+      <p>
+        Formado em Sistemas de Informação em 2019, tenho na área de qualidade de
+        software desde 2013. Recentemente conclui a Especialização em Engenharia
+        de Softeware, em 2022, pela UNISINOS/RS, e atuando como Engenheiro de
+        Qualidade no Sicredi. Gosto de aprender novos conhecimentos e realizar
+        experimentos com novas tecnologias.
+      </p>
+    </section>
+  );
+}
