+++ src/components/profile-info/profile-info.module.css
@@ -1,0 +1,6 @@
+.headingMd {
+    font-size: 1.2rem;
+    line-height: 1.5;
+    text-align: justify;
+    text-justify: inter-word;
+}
